requirements.txt
asgiref==3.7.2
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
Pygments==2.15.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.2
